+++ app/workers/pusher_notification_worker.rb
@@ -1,0 +1,8 @@
+class PusherNotificationWorker
+  include Sidekiq::Worker
+
+  def perform(channel_name)
+    Pusher.trigger(channel_name, 'notify_changes',
+      message: I18n.t('update_stories_successfully'))
+  end
+end