+++ config/webpack/shared.js
@@ -12,6 +12,19 @@ const ManifestPlugin = require('webpack-manifest-plugin')
 const extname = require('path-complete-extname')
 const { env, settings, output, loadersDir } = require('./configuration.js')
 
+const dotenv = require('dotenv')
+
+const dotenvFiles = [
+  `.env.${env.NODE_ENV}.local`,
+  '.env.local',
+  `.env.${env.NODE_ENV}`,
+  '.env'
+]
+
+dotenvFiles.forEach((dotenvFile) => {
+  dotenv.config({ path: dotenvFile, silent: true })
+})
+
 const extensionGlob = `**/*{${settings.extensions.join(',')}}*`
 const entryPath = join(settings.source_path, settings.source_entry_path)
 const packPaths = sync(join(entryPath, extensionGlob))